+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2024-04-17
+
 ### Fixed
 - Added `TargetOrSelector` type alias to typedoc.
 
@@ -90,7 +92,8 @@ Thanks to [klasjersevi](https://github.com/klasjersevi)
 ## [0.2.0] - 2022-06-10
 First release tag.
 
-[Unreleased]: https://github.com/nicholas-wilcox/lit-modal-portal/compare/v0.6.2...HEAD
+[Unreleased]: https://github.com/nicholas-wilcox/lit-modal-portal/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/nicholas-wilcox/lit-modal-portal/compare/v0.6.2...v0.7.0
 [0.6.2]: https://github.com/nicholas-wilcox/lit-modal-portal/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/nicholas-wilcox/lit-modal-portal/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/nicholas-wilcox/lit-modal-portal/compare/v0.5.0-pre...v0.6.0